requirements.txt
requests==2.18.4
overpy==0.4
numpy==1.14.1
# tensorflow-gpu==1.5.0
# tensorflow==1.5.0
keras==2.1.4
scipy==1.0.0
scikit-image==0.13.1
pillow==5.0.0
h5py==2.7.1
opencv-python==3.4.0.12
pyGeoTile==1.0.5
shapely==1.6.4.post1
django==2.0.2
djangorestframework==3.7.7
airtiler==0.4.0
jupyter_contrib_nbextensions
widgetsnbextension
ipywidgets
matplotlib
Cython
pycocotools
